Рефактор main-page
