+++ src/features/map/hooks/use-current-city-coord.ts
@@ -1,6 +1,6 @@
-import type { City, Points } from '@/widgets/map/types';
+import type { City, Points } from '@/features/map/types';
 import useFilteredOffersByCity from '@/entities/Offers/hooks/use-filtered-offers-by-city';
-import { mapPointMapper, mapCityMapper } from './../utils/map-point-mapper';
+import { mapPointMapper, mapCityMapper } from '../utils/map-point-mapper';
 import { CITY } from '@/mocks/city';
 import { POINTS } from '@/mocks/points';
