+++ src/pages/offer-page.tsx
@@ -3,8 +3,8 @@ import CommentsList from '@/entities/Comments/components/comments-list/comments-
 
 import OffersCardList from '@/entities/Offers/components/offers-card-list/offers-card-list';
 
-import Map from '@/widgets/map/map';
-import { mapPointMapper } from '@/widgets/map/utils/map-point-mapper';
+import Map from '@/features/map/map';
+import { mapPointMapper } from '@/features/map/utils/map-point-mapper';
 
 import { reviews } from '@/mocks/reviews';
 import { CITY } from '@/mocks/city';
